my_button_code
Function: Use BUTTON1 and BUTTON2 to light the PURPLE and BLUE LED. When
the button is currently pressed,  make sure the right LED is on.And use
BUTTON1 to control the RED LED is on or off.When the RED LED is on,
start blinking the LED at the current rate.
And when you pressed BUTTON3, the blinking rate will change.